Message should be encode special xml chars
The message should be encoded when it gets transformed to an XML node. We are using this library to log errors to a chat room. Errors can contain all kinds of characters, including reserved ones like < and >. This fixes the encoding of the message.

Update Message.php

Update Message.php

Update Message.php